pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wiki"
version = "1.0.0"
description = "Karpathy-style LLM-driven living knowledge base"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core
    "httpx>=0.27",
    "pyyaml>=6.0",
    "watchfiles>=0.21",
    "jinja2>=3.1",

    # Document parsing (Docling preferred, fallbacks below)
    "docling>=2.5.0",

    # Docling fallbacks
    "pymupdf>=1.24",          # fitz — fast PDF text
    "pypdf>=4.0",             # pure-python PDF fallback
    "python-docx>=1.1",       # .docx parsing
    "pytesseract>=0.3.10",    # image OCR
    "Pillow>=10.0",

    # Wiki / output rendering
    "python-frontmatter>=1.1",
    "markdown>=3.6",

    # Charts
    "matplotlib>=3.8",

    # Async
    "anyio>=4.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.2",
    "langchain-community>=0.2",
]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]
all = ["llm-wiki[langchain,server]"]

[project.scripts]
llm-wiki = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["wiki_engine*", "cli*"]
